Simplify Address changeset (#99)
Motivation

Pivotal story: #158236559
The address changeset function could not be used in a cast_assoc/3, hence the need to refactor.

Description

Renamed create_changeset/4 to changset/2

The changeset function no longer accepts state and country structs. This saves unnecessary complication in their validation. changeset/2 can be used for both creation and update!
Moreover, when controllers will invoke the function for creation or update, they would not be able to "load" state and country structs (it is also not their responsibility).
The validation logic is also much more readable.

[delivers #158236559]