Manage Permissions feature (#115)
The backend of an E-commerce system is typically
managed by an admin and a team comprising of users having varied
degree of access to the resources. To achieve this, the
admin needs to create a set of roles and associate these roles
with various permissions to create different access levels.

This change
- Adds a permission table in the database.
- Adds a model layer for permissions to do the
  usual CRUD functions. UI interface in the admin
  application was created for the same.
- The role functionality was modified a little
  to accomodate associating permissions while
  adding new roles.
- Addition of role_manifest.yaml allows pemissions
  in the system to be associated with controllers and function.
- Access to a particular resource is now handled by a
  controller plug which verifies whether a user can access
  the resource or not.
- [delivers #158439773]