Embedding addresses in order (#104)
* Compute and persist shipment when provided addresses

The `add_addresses` event performs:
* persists the addresses
* computes default packages from each location
* finds an optimal shipment
* splits any "too heavy" packages into smaller ship-able ones.

The shipment is available in the `Context.state.shipment` field, and
could be an empty list - indicating that there is no ship-able
fulfillment for this order.

Persist shipment to DB
----------------------

This change inserts packages and their items together in a DB
transaction using the new `Domain.Shipment.to_package/2` and the
`Model.Package.create/1`.

Change to `EmbeddedShippingMethod`
----------------------------------

It is not possible to use the Money struct in an embed. Need to open an
issue upstream.

Misc changes
------------
1. Remove `:number` from `PackageItem` schema.
2. Add test for `to_package/2`

[delivers #157926007]

* Short circuit a few computations

This change saves us from a few errors when a shipment is empty but
somewhere there's a reduce that tries to consume it.

* Add changeset for (partial) order update

* Simplified `add_address` event, the whole thing runs in a multi now!

* Add tests for order transition functions

This change simplifies the flow for associate address. It is important
to keep this transition impure.

Misc
----
* Remove defunct `model/stock.ex`
* Remove the unused `create/2` clause in `Model.Package`

* Add number field to PackageItem

The field was removed earlier (in this branch) but we need it back as it
is useful.

Improved tests for `PackageItem` as `Shipment.to_package/2`

[delivers #158236509]

* Fix shipment testswq

* embeded addresses in order

* added migration files

* removed address_ids and updated order schema

* test cases updated for emebedding address in order

* Implemented embedded schema for order address

* patching up the tests after variants seeded

* updated code readability issues

* updated credo related issues

* removed `map_from_struct` function

* Update mix.lock and corrected seed

* Remove Package has_one address through order

Since shipping_address is an embed in order, we can't make ecto load the
embed via "has_one through" in package.