Minor Fix Package Item Cascade delete
Prior to this change,
On deleting a line item if package items were in the system for, it
was raising an error.

This change
Allows deleting a lineitem if packages exist. Even if packages have been
created for a lineitem of user deletes the lineitem, package item should
be removed.

- Added a migration to cascade delete package items for a lineitem in
 an order.