Taxon (category) deletion (#367)
Why?
User should be able to delete a particular category from the taxonomy.

This change addresses the need by:
Adding an option to delete a particular taxon. This also delete the associated products with the taxon.
If a taxon has multiple child taxons then those also get deleted along with their associated products. 

[delivers #161198500]