Order number format modification (#415)
Why?
At present, we are using many different special characters to generate the order number. This does not result in a well readable format of the order number. It also leads to communication issues and difficulties with customer support. 

We should generate human readable and communication friendly order number format. 

This change addresses the need by:
Autogenerating order number of 17 digits made by permutation of the characters in a given string -
"0123456789-"

[finishes #162498460]