[feature] add tax configuration and tax class (#440)
Why?
Tax configuration is required to store general configuration data for
handling taxes. Tax Classes are used around the system as templates
which would be set for a tax rate and would assist in tax calculation.

-> Tax Configuration
It has been handled using `single row` modelling wherein data is stored in a single row with all the attributes of configurations as columns. 
An EAV approach was taken earlier to handle this but it has some limitations which are mentioned in 
detail [here](https://www.pivotaltracker.com/story/show/163364131).

-> Note
The older method of handling tax through `tax_categories` and `tax_rates` has been removed in this PR.

This change addresses the need by:
- Adding schema, migration for tax class and tax configuration.
- Adding CRUD API for the above entities.
- Adding seed for setting up initial configuration data.

[delivers #163283435]
[delivers #163282965]