Handle error on duplicating product name (#477)
## Why?
On entering the same name for different products error was being raised as unique constraint didn't exist a schema level.
Also, at DB level unique constraint existed only for the `slug`. This uniqueness constraint should be applied only to the products which are not deleted in the system. Hence, uniqueness is now checked on a combination of `slug` and `deleted_at` product field.

## This change addresses the need by:
- Created a unique_index on combination of `slug` and `deleted_at` in migrations.
- Added unique_constraint on slug and deleted_at in product schema.
- Deleted_at is being handled using unix_timesamp at the schema level.

[delivers #164069814]